backend/requirements-prod.txt
# Production Requirements for Cloud Run Deployment
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
firebase-admin==7.1.0

# Additional production optimizations
gunicorn==21.2.0

# Security and performance
email-validator==2.1.1
pydantic[email]==2.9.2
